feat: add support for additional texture(s) in raymarching shaders.

+++ core/src/gaiasky/scene/system/initialize/BackgroundModelInitializer.java
@@ -42,7 +42,7 @@ public class BackgroundModelInitializer extends AbstractInitSystem {
         var renderType = Mapper.renderType.get(entity);
 
         // Size
-        body.size *= Constants.DISTANCE_SCALE_FACTOR;
+        body.size *= (float) Constants.DISTANCE_SCALE_FACTOR;
 
         // Force texture loading
         model.renderConsumer = ModelEntityRenderSystem::renderGenericModel;
